+++ CHANGELOG.md
@@ -5,9 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+### Added
+
  - Add `pyidide minify` command to minify the Python packages with AST rewrites by,
    removing comments and docstrings
    [#23](https://github.com/pyodide/pyodide-pack/pull/23)
+
+### Changed
+
+ - Added support for Pyodide 0.24.0. This is now the minimal supported version of Pyodide.
+   [#26](https://github.com/pyodide/pyodide-pack/pull/26)
 
 ## [0.2.0] - 2022-09-04